cleanup(angular): cleanup remote generator
