+++ packages/angular/src/generators/setup-mf/lib/index.ts
@@ -9,3 +9,4 @@ export * from './set-tsconfig-target';
 export * from './setup-host-if-dynamic';
 export * from './setup-serve-target';
 export * from './update-host-app-routes';
+export * from './remove-dead-code-from-remote';
